requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap4==2.3.1
idna==2.10
pycryptodome==3.6.6
pyotp==2.6.0
pypng==0.0.20
PyQRCode==1.2.1
python-firebase @ git+https://github.com/ozgur/python-firebase@0d79d7609844569ea1cec4ac71cb9038e834c355
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
